+++ storage_class/config.go
@@ -8,12 +8,15 @@ import (
 	"github.com/netapp/trident/storage_attribute"
 )
 
+// UnmarshalJSON parses a JSON-formatted byte array into a storage class config struct.
 func (c *Config) UnmarshalJSON(data []byte) error {
 	var tmp struct {
-		Version             string              `json:"version"`
-		Name                string              `json:"name"`
-		Attributes          json.RawMessage     `json:"attributes,omitempty"`
-		BackendStoragePools map[string][]string `json:"requiredStorage,omitempty"`
+		Version         string              `json:"version"`
+		Name            string              `json:"name"`
+		Attributes      json.RawMessage     `json:"attributes,omitempty"`
+		Pools           map[string][]string `json:"storagePools,omitempty"`
+		RequiredStorage map[string][]string `json:"requiredStorage,omitempty"`
+		AdditionalPools map[string][]string `json:"additionalStoragePools,omitempty"`
 	}
 	err := json.Unmarshal(data, &tmp)
 	if err != nil {
@@ -22,20 +25,31 @@ func (c *Config) UnmarshalJSON(data []byte) error {
 	c.Version = tmp.Version
 	c.Name = tmp.Name
 	c.Attributes, err = storage_attribute.UnmarshalRequestMap(tmp.Attributes)
-	c.BackendStoragePools = tmp.BackendStoragePools
+	c.Pools = tmp.Pools
+
+	// Handle the renaming of "requiredStorage" to "additionalStoragePools"
+	if tmp.RequiredStorage != nil && tmp.AdditionalPools == nil {
+		c.AdditionalPools = tmp.RequiredStorage
+	} else {
+		c.AdditionalPools = tmp.AdditionalPools
+	}
+
 	return err
 }
 
+// MarshalJSON emits a storage class config struct as a JSON-formatted byte array.
 func (c *Config) MarshalJSON() ([]byte, error) {
 	var tmp struct {
-		Version             string              `json:"version"`
-		Name                string              `json:"name"`
-		Attributes          json.RawMessage     `json:"attributes,omitempty"`
-		BackendStoragePools map[string][]string `json:"requiredStorage,omitempty"`
+		Version         string              `json:"version"`
+		Name            string              `json:"name"`
+		Attributes      json.RawMessage     `json:"attributes,omitempty"`
+		Pools           map[string][]string `json:"storagePools,omitempty"`
+		AdditionalPools map[string][]string `json:"additionalStoragePools,omitempty"`
 	}
 	tmp.Version = c.Version
 	tmp.Name = c.Name
-	tmp.BackendStoragePools = c.BackendStoragePools
+	tmp.Pools = c.Pools
+	tmp.AdditionalPools = c.AdditionalPools
 	attrs, err := storage_attribute.MarshalRequestMap(c.Attributes)
 	if err != nil {
 		return nil, err